test(release): split release note tests
