+++ tools/release-notes/test_helpers_test.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"strings"
+	"testing"
+)
+
+func assertContains(t *testing.T, haystack, needle string) {
+	t.Helper()
+	if !strings.Contains(haystack, needle) {
+		t.Fatalf("expected output to contain %q:\n%s", needle, haystack)
+	}
+}
